refactor `FeeParameters` to parent transfer module
